+++ mdxcanvas/xml_processing/xml_processing.py
@@ -9,7 +9,7 @@ from .quarto_slides_preprocessor import make_quarto_slides_preprocessor
 from .syllabus_tags import SyllabusTagProcessor
 from ..resources import ResourceManager
 from ..util import parse_soup_from_xml
-from ..xml_processing.group_tags import AssignmentGroupTagProcessor
+from ..xml_processing.group_tags import AssignmentGroupTagProcessor, GroupCategoryTagProcessor
 from ..xml_processing.module_tags import ModuleTagProcessor
 from ..xml_processing.page_tags import PageTagProcessor
 from ..xml_processing.quiz_tags import QuizTagProcessor
@@ -80,6 +80,7 @@ def process_canvas_xml(resources: ResourceManager, text: str):
         'announcement': AnnouncementTagProcessor(resources),
         'assignment': AssignmentTagProcessor(resources),
         'group': AssignmentGroupTagProcessor(resources),
+        'group-category': GroupCategoryTagProcessor(resources),
         'module': ModuleTagProcessor(resources),
         'navigation': NavigationTagProcessor(resources),
         'page': PageTagProcessor(resources),